requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
aiosqlite==0.22.1
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
appdirs==1.4.4
attrs==26.1.0
banks==2.4.1
brotli==1.2.0
certifi==2026.2.25
charset-normalizer==3.4.6
click==8.3.1
colorama==0.4.6
contourpy==1.3.3
cycler==0.12.1
dataclasses-json==0.6.7
datasets==4.8.4
Deprecated==1.3.1
dill==0.4.1
dirtyjson==1.0.8
diskcache==5.6.3
distro==1.9.0
docstring_parser==0.17.0
faiss==1.8.0
fastapi==0.135.2
ffmpy==1.0.0
filelock==3.25.2
filetype==1.2.0
fonttools==4.62.1
frozenlist==1.8.0
fsspec==2026.2.0
gradio==6.10.0
gradio_client==2.4.0
greenlet==3.3.2
griffe==2.0.2
griffecli==2.0.2
griffelib==2.0.2
groovy==0.1.2
h11==0.16.0
hf-gradio==0.3.0
hf-xet==1.4.2
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
huggingface_hub==1.8.0
idna==3.11
instructor==1.14.5
Jinja2==3.1.6
jiter==0.11.1
joblib==1.5.3
jsonpatch==1.33
jsonpointer==3.1.1
kiwisolver==1.5.0
langchain==1.2.13
langchain-classic==1.0.3
langchain-community==0.4.1
langchain-core==1.2.22
langchain-experimental==0.4.1
langchain-huggingface==1.2.1
langchain-ollama==1.0.1
langchain-openai==1.1.12
langchain-text-splitters==1.1.1
langgraph==1.1.3
langgraph-checkpoint==4.0.1
langgraph-prebuilt==1.0.8
langgraph-sdk==0.3.12
langsmith==0.7.22
llama-index==0.14.19
llama-index-cli==0.5.6
llama-index-core==0.14.19
llama-index-embeddings-openai==0.6.0
llama-index-instrumentation==0.5.0
llama-index-llms-openai==0.7.3
llama-index-workflows==2.17.1
markdown-it-py==4.0.0
MarkupSafe==3.0.3
marshmallow==3.26.2
matplotlib==3.10.8
mdurl==0.1.2
mpmath==1.3.0
multidict==6.7.1
multiprocess==0.70.19
mypy_extensions==1.1.0
nest-asyncio==1.6.0
networkx==3.6.1
nltk==3.9.4
numpy @ file:///Users/runner/miniforge3/conda-bld/numpy_1707225640867/work/dist/numpy-1.26.4-cp311-cp311-macosx_11_0_arm64.whl#sha256=b62b8b824024855c3eb209bd45bfcf60bc9a11b94cbf9d5b7b25dc503021e672
ollama==0.6.1
openai==2.30.0
orjson==3.11.7
ormsgpack==1.12.2
packaging @ file:///opt/miniconda3/conda-bld/packaging_1761049079066/work
pandas==3.0.1
pillow==12.1.1
pinecone-client==6.0.0
pinecone-plugin-interface==0.0.7
platformdirs==4.9.4
propcache==0.4.1
pyarrow==23.0.1
pydantic==2.12.5
pydantic-settings==2.13.1
pydantic_core==2.41.5
pydub==0.25.1
Pygments==2.19.2
pyparsing==3.3.2
pypdf==6.9.2
pyrate-limiter==4.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-multipart==0.0.22
pytz==2026.1.post1
PyYAML==6.0.3
ragas==0.4.3
regex==2026.2.28
requests==2.33.0
requests-toolbelt==1.0.0
rich==14.3.3
safehttpx==0.1.7
safetensors==0.7.0
scikit-learn==1.8.0
scikit-network==0.33.5
scipy==1.17.1
sec-edgar-downloader==5.1.0
semantic-version==2.10.0
sentence-transformers==5.3.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.48
starlette==0.52.1
sympy==1.14.0
tenacity==9.1.4
threadpoolctl==3.6.0
tiktoken==0.12.0
tinytag==2.2.1
tokenizers==0.22.2
tomlkit==0.13.3
torch==2.11.0
tqdm==4.67.3
transformers==5.4.0
typer==0.24.1
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uuid_utils==0.14.1
uvicorn==0.42.0
wrapt==2.1.2
xxhash==3.6.0
yarl==1.23.0
zstandard==0.25.0
